rename to image-watcher
I realized that the program which generates the png and the program
which displays the png don't need to be the same program. In particular,
I can use "ghcid --test=..." to run an IO action which updates a png
file every time my Haskell program successfully compiles, and now all I
need is a program which watches that png for changes and displays the
latest version.